+++ app/views/users/show.html.erb
@@ -1,4 +1,3 @@
-
 <%= render 'archive_banner' if @archive.present? %>
 
 <div class="container text-dark-blue h4-as-labels mt-4">
